feat: 결과를 보고 싶은 사람은 한 명 이상이어야 한다
