+++ README.md
@@ -28,8 +28,8 @@
   - [x] 사다리 높이는 숫자이어야한다.
   - [x] 사다리 높이는 1 이상이어야한다.
 - [ ] 결과를 보고 싶은 사람을 입력받는다.
-  - [ ] 참여한 사람이어야 한다.
-  - [ ] 한 명 이상이어야 한다.
+  - [x] 참여한 사람이어야 한다.
+  - [x] 한 명 이상이어야 한다.
   - [ ] all 를 입력하면 전체 참여자의 실행 결과를 출력한다.
   - [ ] all or 없음 을 입력하면 실행 결과 출력 후 종료된다.
 - [x] 사용자가 잘못된 입력을 하면 해당 부분부터 다시 입력받는다.
